Replaces uses of Fixnum with llvm::PointerEmbeddedInt.
The two types are nearly identical, and Fixnum is only in the Swift branches of LLVM,
not in mainline LLVM. The main change here is that PointerEmbeddedInt's representation
is always a pointer, which means sizeof and anything depending on it give you the size
of a pointer rather than the smallest primitive integral type that can hold the number.
But for serialization purposes being explicit is probably a good thing, although it
does mean more coordination across files.