[Swift] Extend swift_name attribute to describe members.
Extend the syntax of the swift_name string to allow "Foo.bar", which
specifies that the entity to which the attribute appertains should be
imported as a member named "foo" of the type with the Swift name
"Foo".

For functions that will become instance methods, the special argument
name "self" indicates which parameter of the C function will become
the "self" parameter. For example, given:

  typedef struct __attribute__(((swift_name("Foo"))) AAFoo {
    /* ... */
  } AAFoo;

  void AAFooDoSomething(AAFoo foo, double x, double y)
    __attribute__((swift_name("Foo.doSomething(self:x:y:)")));

The C type AAFoo will be imported as the Swift type Foo, and
AAFooDoSomething will be imported as an instance method
"doSomething(x:y:)" of Foo, where the first parameter is "self".

Part of implementing SE-0033.