[Swift] Extend swift_name to support getters and setters.
Extend the string format of swift_name to support labeling a function
as a getter or setter for a given property name, so that the getter
(and optional setter) function(s) will be imported into Swift as a
property. Slightly extended from what SE-0033 actually requires, but
useful in its own right.