[Parser] Fix look ahead after EOF while parsing objc message and lambdas
If a closing ')' isn't found for a macro instantiation inside a '[',
the next token is EOF, this leads to crashes if we try to look ahead of
that. This could be triggered whenever trying to parse lambdas or objs
message expressions.

Differential Revision: http://reviews.llvm.org/D20451

rdar://problem/25662647

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@271314 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 9d98619f8ed131a9363f39ae0eb07a431e4d6c42)