Reverting part of r276455 Add .rgba syntax extension to ext_vector_type types
Actually revert the addition of the new syntax. The patch was doing 2 things:
 - adding the new rgba syntax
 - makes it mandatory to use the 's' prefix for the indexing with hex indexing

The issue is that v.a becomes ambigous. It was accepted as accessing the 10th
index before, and now it means the 3rd component. For one release, force people
to use the s prefix ('v.sa') and do not add the rgba syntax to prevent any
ambiguity in the code.

rdar://problem/28125882