Fix PR 28885: Fix AST Printer output for the inherited constructor using declarations.
This commit ensures that the correct record type is printed out for the
using declarations that represent C++ inherited constructors.
It fixes a regression introduced in r274049 which changed the name that's
stored in the using declarations that correspond to inherited constructors.

Differential Revision: https://reviews.llvm.org/D25131

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@283105 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 549741a3bcb6006a7afb39f048f7a202c35e66ae)