[AST] Remove superfluous docblock change
Upstream Clang documents the `Message` parameter as simply `\param`,
while Swift's fork of Clang documents it as `\param[out]`. Although
`\param[out]` may be more correct, the change to the documentation
should be made in upstream Clang, not Swift Clang.

Furthermore, upstream Clang does not appear to have a strong convention
as to the use of `\param[out]`: of 4616 instances of `\param`, only 49
are denoted as `\param[out]`.