Fix merge problem from clang r288295.
John added the IsForDefinition argument to
CGObjCNonFragileABIMac::GetClassGlobal on trunk. This merged badly with
the Swift version of the code that previously had a boolean argument
named "ForDefinition". In the new version "ForDefinition" is an enum value
so this code did not fail to compile and only caused a test failure.
rdar://problem/29538073